requirements.txt
fastapi>=0.111
uvicorn>=0.29
supabase>=2.4
langchain>=0.2
langchain-huggingface>=0.0.6
langchain-core>=0.2
langgraph>=0.1
pydantic>=2.0
python-dotenv>=1.0
PyGithub>=2.3
python-multipart>=0.0.9